chore: Refactor wasm detection
