+++ src/Uno.DevTools.Telemetry/TelemetryCommonProperties.cs
@@ -28,20 +28,7 @@ namespace Uno.DevTools.Telemetry
 {
     internal sealed class TelemetryCommonProperties
     {
-        private static readonly bool IsWasmBrowser = DetectWasmBrowser();
-
-        private static bool DetectWasmBrowser()
-        {
-#if NET5_0_OR_GREATER
-            return OperatingSystem.IsBrowser() || OperatingSystem.IsWasi();
-#else
-            // For netstandard2.0, check RuntimeInformation.OSDescription
-            var osDescription = RuntimeInformation.OSDescription;
-            return osDescription.Contains("Browser", StringComparison.OrdinalIgnoreCase) ||
-                   osDescription.Contains("WebAssembly", StringComparison.OrdinalIgnoreCase) ||
-                   osDescription.Contains("WASI", StringComparison.OrdinalIgnoreCase);
-#endif
-        }
+        private static readonly bool IsWasmBrowser = PlatformDetection.IsWasmBrowser;
 
         public TelemetryCommonProperties(
             string storageDirectoryPath,
